Map & Traps Prototype completed.
No Player input yet, just move the testsubject (blue) manually. Purple is Push Trap, it push to the end of road its facing (Default is top). the brown is map swap trap it swaps two section (one its on and one its facing ) (Default is top). Gray tiles are wall, Whites are floor, and light green is the Goal!